pyproject.toml
[project]
name = "researchbook"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-search-results>=2.4.2",
    "matplotlib>=3.10.3",
    "networkx>=3.5",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
    "scholarly>=1.7.11",
    "streamlit>=1.45.1",
]
